backend/requirements.txt
Django==4.2.7
djangorestframework
pandas
reportlab
djangorestframework-simplejwt
gunicorn
whitenoise
django-cors-headers